tests/wpt/web-platform-tests/workers/WorkerGlobalScope_close.htm

<!DOCTYPE html>
<title> WorkerGlobalScope close(): clear events queue </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker('./support/WorkerClose.js');
  worker.onmessage = t.step_func(function(e) {
    assert_equals(e.data, "ping");
    worker.onmessage = t.unreached_func("Unexpected message event");
    worker.postMessage("pong");
    setTimeout(t.step_func_done(), 100);
  });
  worker.postMessage("ping");
});
</script>
